Migrate dunst from iniparser to inih.
Instead of using iniparser, which is rarely packaged, and is an
external dependency, use an embedded copy of inih instead, which is
only a hundred or so lines, and can do pretty much the same as
iniparser.

The benefit is one less external dependency, and it makes it easier
for distributions to package dunst.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>