Use XKeysymToKeycode
When attempting to define a shortcut (other than ctrl+shift+space) that
uses shift, I would receive a warning (Unknown keyboard shortcut) and
the shortcut would not function. Replacing the search for the proper
keycode with a call to XKeysymToKeycode allows me to define and
correctly use a shortcut involving shift.