change index for null-termination of dmenu output
The index used to null-terminate the buffer containing the output of
dmenu has been changed from len-1 to len: before this change the last
character of the url had been ripped off. The access to this character
is legitimate because the buffer is preallocate to 1024 and data from
read are truncated at 1023.