option_parser.c: fix load_ini_file
The bit of the parser that dealt with quoting and comments was broken.  I
have fixed it by having it crawl (-ish; using strbrk) the string for
characters that it treats specially, toggling whether it is in a quoted
portion via a "in_quote" variable.

This means that it is not able to "jump" over quoted strings; it must find
each "#" and ";" in them, then realize that it is in a quoted string, and
needs to ignore comment characters.  This is of course slower than being
able to skip over the quote, but I think it is more readable.  It should
also be easier to extend (perhaps supporting backslash escapes?).  Anyway,
at this level, the performance changes don't matter anyway (premature
optimization).