Chain history_ignore and transient together
If the transient hint is set on the notification, it should only be
valid as long as it shows up. So the pushing it to the history
makes no sense.

As it's no problem to change this individually by the rule system,
it's a sane default for history_ignore to default to the transient
option.