Refactor x11 event source to handle multiple windows
Move the code to register the x11 event source under x_win_create so
multiple windows can be created and be properly registered as the event
source.

One more missing piece is that the callback functions are still using
the global win variable and not getting the window from the source.