Do not search absolute paths in the icon_path
The icon_path has no point for searching an absolute filename. It was
actually meant to avoid the complex lookup algorithm of the icon theme
spec. And the iconname in this case should be actually just the icon's
basename when searching the path.

Concatenating icons like `~/testimg.png` with all icon paths has no
point and also an additional suffix has no point.