Pass NULL instead of empty string to option parser
A bug, introduced in #576. If alignment isn't set in the config, the
settings would print a warning about an invalid value. But actually it's
just the default string, which can't get parsed.

The fallback value should get handled by a false return value and not by
the option_get_string's default value.