Replace set_transient with idle_threshold
Since the definition of transient in the spec is quite ambiguous we
should avoid associating with a behaviour directly. Instead we should
just expose both ends of the chain via rules and let the user join them
together if necessary.

As such the `transient` hint can now only be matched via rules and
idle_threshold can be set to 0 in a rule to bypass the existing setting.