wayland: Return the largest scale when output cannot be found
When follow mode != none dunst cannot detect what output is being used
to display the notification (yet). With this commit dunst falls back to
using the largest scale from any ouput. The compositor should scale the
buffer down again if the scale of the output is smaller that the used
scale.