Icon: make the icon lookup recursive
This code is taken from mako's icon.c and only slightly adapted to use
with dunst. Due to this change the icon_path should be changed to the
root of the icon theme, instead of a subdirectory of it. The icon size
may also be different from before, since the largest icon that still
fits is used, instead of the first icon found. If this icon is too
large you should change max_icon_size accordingly.