config.mk: Better defaults for service units.
Don't install systemd user service and dbus service units in directories
which are managed by the distribution by default. The old checks for
SERVICEDIR* relied on pkg-config which will return the location for
actual distribution packages (i.e. /usr/share/dbus-1/services) in most
cases, unless the user has built systemd and dbus locally, which should
be the exception rather than the rule.

Most users will expect things to go to /usr/local, so use the officially
documented default locations for locally installed software under that
prefix. See Table 2 in systemd.unit(5) and the CONFIGURATION FILE
section in dbus-daemon(1) for reference. This way distro packages and
locally installed software can coexist peacefully with the local version
taking precedence in any sane installation.

Since there is not much to check, given the above, put the defaults in
config.mk.

Also change the executable for automatically detecting systemd. The
check said "systemd", so actually check for that executable and not
systemctl which is not the same. In some obscure case there might be
systemctl but not systemd, maybe. At least there seems to be no good
reason for not checking the existence of systemd.