Less disruptive (un)install.
Check if the destination dunstrc exists and if so skip installation
unless forced by SYSCONF_FORCE_NEW. Also don't remove it on uninstall.
There is a new target uninstall-purge (think: apt-get purge) for getting
rid of everything, that was actually installed. Don't just blindly and
*forcibly* remove SYSCONFDIR, there might be backups of configs in
there, or whatever; just be nicer to the user/admin. ;)

New (un)install-dunstrc targets. It is an rc, after all, and not just
some doc item. And it actually works now in every environment. There are
some warnings about non-existent (obsolete?) settings, though. Leave
those alone for now, since they are out of scope for this task.

New location for SYSCONFDIR. Usually locally installed packages have
their own system configs under /usr/local/etc. Make sure dunstrc finds
its proper place there and that it gets used exclusively to avoid
conflicts or other unexpected behaviour if a distro version of dunst is
also present, which previously took precedence.

Also expose a new optional make parameter for the full path to dunstrc.
It gets used multiple times in Makefile, so why not exose it in
config.mk for builders who need more flexibility?

New CPP directive defines SYSCONFFILE location for use in settings.c.
There is no point in trying other locations for the system-wide dunstrc.
It goes in exactly one location and that is known at compile time, so
make use of that fact and get rid of some LOCs in the process.

test/test-install.sh: Test the new install targets
Also restructured the script to make it more concise and reduce
repetition. Plus, it is no longer necessary to manually sort the list of
files that are expected after install.