log.h: Refine prefix of some message levels.
Dial back a little on the space requirement of the prefix for messages
of levels "debug", "error" and "critical". The prefix now only
includes the function name, padded to 16 characters so as to make for
better vertical alignment, and the line number in the source file.

Additionally updated feature test macro to make this work with recent
enough versions of clang as well.

config.mk: Disable warnings about empty __VA_ARGS__

Since compilation is done with '-std=gnu99' and token pasting of ',' and
'__VA_ARGS__' is a GNU extension which also works with clang (>=6) this
warning should be safe to disable, see also the changes to log.h above.
The CI did not complain in my local tests.

Only clang threw this warning before and only to inform that it is
indeed a GNU extension, despite using the same '-std=gnu99'.
This should not mask genuine errors when __VA_ARGS__ must not be empty,
i.e. usage without token pasting it with ',', since those should be
errors that lead to unsuccessful compilation, anyway.