icon: scale icons from the icon theme to icon_size
This makes a difference when the icon of a theme is used for multiple
sizes. These icons can be scalable or not, but either way, the
index.theme states that these icons can be scaled to this size.

When using fractional scaling, icons may look blurry, depending on your
icon theme.