Add support for notification gaps
To support, two modifications were required.

The first increases the size of the total image surface used by cairo so
that the extra space required is not cropped. To calculate this, the gap
size plus size of top and bottom border frames are summed for each
notification. This is because normally only two frame widths are shown,
but with gaps they are shown for each notification.

The second simply increments the y offset by an additional value of the
gap size. Each notification is also marked as both first and last to
force both top and bottom frame borders to be drawn (if enabled).