icon: revert back to using min_icon_size and max_icon_size
The recursive icon lookup used an icon_size rule to scale all icons to
this size, but this was a bit limiting in allowing bigger icons for
things like album art. To fix this, the old behaviour has been
reinstated. The main difference is that icons from a theme (that are
searched by icon name) are of the size `min_icon_size`, instead of being
of whatever size was found first.

ref: #1037