Fix raw icons not being scaled according to icon size
This bug was caused by max_icon_size and min_icon_size being a rule,
thus the value being known later (only after notification_init()) and
new_icon needing to have the icon available in that same function to
override it. This issue is solved with a boolean that signals if the
icon is overridden.

Fixes #1043