High CPU usage when show_age_threshold = -1
The show_age_threshold can be set to -1 to disable the feature that
updates notifications with their age.

In this scenario the queues_get_next_datachange may return -1 to
indicate that there is no next wakeup time (e.g. when there are
permanent / critical notifications).

The code that calls queues_get_next_datachange should take this
scenario into account and not schedule another timeout if -1 is
returned.