Wayland: suport GTK/CSS cursor names
Adwaita, the default icon theme for many distributions, decided to stop
providing legacy Xcursor icons in favor of the GTK/CSS names[1].
This requires us to probe for "default" instead of the "left_ptr", as
the latter is no longer available in Adwaita.
Keep the fallback to "left_ptr" for more conservative themes.

[1]: https://gitlab.gnome.org/GNOME/adwaita-icon-theme/-/commit/74e9b79471236320d2af4925d6c5bb7df22380ce