makefile: add version variable quotes in git case
Add quotes to the VERSION variable, when they are determined by invoking
git. While git describe doesn't seem to output spaces. But the fallback
case of just echoing the VERSION variable may contain spaces, as every
release includes a space separated release date in the version variable.

This fixes the issue that if the .git directory is present on a release,
but the git command isn't available or fails, the VERSION variable isn't
escaped anymore and the release date contained causes a build failure.