wayland: improve deinit and init
Improve the wayland deinit and init process to fix problems after
running deinit followed by init in case of a reload. Furthermore destroy
otherwise leaked objects and clear associated pointers to prevent
potential use after free bugs.

This change fixes dunst from breaking sometimes, when a reload is
triggered while a notification is displayed. This issue happed because
of either ctx.dirty not being cleared, which causes an set_dirty to do
an early exit without scheduling a frame to the surface. Or because
the frame_callback was not being cleared, which prevented further
surface commits, as the callback would never be triggered after the
surface was destroyed by the deinit process.

Fixes #1434