add .editorconfig file
Add an .editorconfig file to automatically configure supported editors.
This file was generated merely based on the existent styles and the
warnings caused while checking the code with the editorconfig-checker
tool.

As the editorconfig checker tool generates false positives when an
indentation is adjusted to match multiline arguments or similar
aesthetic adaptions, the corresponding check was disabled. This will
miss wrong indentation sizes (e.g. 4 instead of 8 spaces).